requirements.txt
APScheduler==3.9.1
flask==2.1.3
psycopg2==2.9.3
websocket-client==1.3.3
websockets==10.3
